+++ data/lightpolution-collector/src/main/java/com/ssafy/countingstar/service/LightPollutionDownloadRecordService.java
@@ -8,13 +8,13 @@ public interface LightPollutionDownloadRecordService {
 	 * @param cksum
 	 * @return 해당 체크섬 데이터가 다운로드 되었는지 확인한다.
 	 */
-    boolean isDownloaded(int cksum);
+    boolean isDownloaded(long cksum);
     
     /**
      * 체크섬을 받아. 해당 체크섬이 다운로드 되었고, 처리되었음을 표시한다.
      * @param cksum
      * @param downloadedTime
      */
-    void markAsDownloaded(int cksum, LocalDateTime date);
+    void markAsDownloaded(long cksum, LocalDateTime date);
 
 }
